requirements.txt
pandas==1.0.0
psycopg2-binary==2.8.4